# Convert folds to take two arguments