@Bodigrim 2021-01-13
I can certainly reproduce your benchmarks with roughly the same results.

Before:
```
folds/foldl'/1                           mean 9.508 ns  ( +- 202.4 ps  )
folds/foldl'/2                           mean 12.66 ns  ( +- 312.4 ps  )
folds/foldl'/4                           mean 18.88 ns  ( +- 608.7 ps  )
folds/foldl'/8                           mean 33.46 ns  ( +- 951.9 ps  )
folds/foldl'/16                          mean 62.70 ns  ( +- 2.429 ns  )
folds/foldl'/32                          mean 111.7 ns  ( +- 3.313 ns  )
folds/foldl'/64                          mean 213.3 ns  ( +- 6.860 ns  )
folds/foldl'/128                         mean 409.9 ns  ( +- 16.32 ns  )
folds/foldl'/256                         mean 814.9 ns  ( +- 27.73 ns  )
folds/foldl'/512                         mean 1.615 μs  ( +- 40.81 ns  )
folds/foldl'/1024                        mean 3.240 μs  ( +- 121.7 ns  )
folds/foldl'/2048                        mean 6.429 μs  ( +- 244.1 ns  )
folds/foldl'/4096                        mean 12.98 μs  ( +- 479.4 ns  )
folds/foldl'/8192                        mean 26.08 μs  ( +- 1.129 μs  )
folds/foldl'/16384                       mean 51.77 μs  ( +- 1.711 μs  )
folds/foldl'/32768                       mean 104.2 μs  ( +- 4.380 μs  )
folds/foldl'/65536                       mean 209.3 μs  ( +- 10.51 μs  )
```

After:
```
folds/foldl'/1                           mean 3.833 ns  ( +- 95.21 ps  )
folds/foldl'/2                           mean 4.195 ns  ( +- 177.2 ps  )
folds/foldl'/4                           mean 6.481 ns  ( +- 161.2 ps  )
folds/foldl'/8                           mean 8.896 ns  ( +- 196.2 ps  )
folds/foldl'/16                          mean 13.77 ns  ( +- 390.6 ps  )
folds/foldl'/32                          mean 27.55 ns  ( +- 476.3 ps  )
folds/foldl'/64                          mean 41.14 ns  ( +- 684.2 ps  )
folds/foldl'/128                         mean 69.15 ns  ( +- 1.425 ns  )
folds/foldl'/256                         mean 124.3 ns  ( +- 2.369 ns  )
folds/foldl'/512                         mean 233.5 ns  ( +- 4.240 ns  )
folds/foldl'/1024                        mean 452.1 ns  ( +- 7.693 ns  )
folds/foldl'/2048                        mean 888.6 ns  ( +- 15.60 ns  )
folds/foldl'/4096                        mean 1.790 μs  ( +- 41.93 ns  )
folds/foldl'/8192                        mean 3.500 μs  ( +- 45.58 ns  )
folds/foldl'/16384                       mean 7.078 μs  ( +- 169.4 ns  )
folds/foldl'/32768                       mean 13.97 μs  ( +- 176.2 ns  )
folds/foldl'/65536                       mean 28.01 μs  ( +- 465.3 ns  )
```