@Boarders 2021-01-13
@Bodigrim I added comments, updated hlint and also I applied the same transformation to similar functions. Every benchmark improves quite a bit as functions are actually getting inlined now.