@Bodigrim 2021-01-14
> Every benchmark improves quite a bit as functions are actually getting inlined now.

Great! Could you please quote results, at least for the functions with pre-existing benchmarks?