@Boarders 2021-01-14
I added your other suggestions too @Bodigrim and have put it all into one commit.