@Bodigrim 2021-01-14
I do not quite understand why `scan{l,r}` and `filter` became so drastically faster. Other examples here (`fold{l,r}{,'}`, `mapAccum{L,R}`) are polymorphic, so inlining monomorphises definitions with understandable performance benefits. But `scan{l,r}` and `filter` are monomorphic already!