@Boarders 2021-01-14
My only guess is that filter/scan getting inlined means that the predicate/accumulator can also get inlined and that can be much better optimized later in the pipeline that core. I really don't know enough to say though.