@Bodigrim 2021-01-14
Yeah, sounds plausible, probably inlining a predicate allows to avoid boxing/unboxing of `Word8`.