@Bodigrim 2021-01-15
What I'm trying to grasp is how much of the speed up can be attributed purely to the deficiencies of eta-reduced expressions used in benchmarks. 

It's probably correct that the most drastic impact is not due to inlining itself, but due to complete elimination of memory allocation. It would be nice to have a robust way to utilise this, not relying on fragile pragams. Is it an evidence to expose a set of functions, taking callbacks of form `Word8# -> Bool`, `Word8# -> Word8# -> Word8#`, etc?