@ethercrow 2021-01-15
> taking callbacks of form Word8# -> Bool, Word8# -> Word8# -> Word8#

These two callbacks can be encoded as a 256 bit mask and a uint8_t[256*256] lookup table respectively and the implementation can be done in C entirely.