@Bodigrim 2021-01-15
@ethercrow yeah, interesting idea. One could have a Template Haskell routine, precomputing masks/lookup tables during compilation and generating the most efficient code for each particular case.

@Boarders could you please check, whether replacing `INLINE` by `INLINEABLE` is sufficient for all functions in this PR?